requirements.txt
greenlet==3.2.2
lmdb==1.6.2
kiwisolver==1.4.7
nvidia-cufile-cu12==1.11.1.6
opt_einsum==3.4.0
setuptools==53.0.0
lightning-utilities==0.14.3
charset-normalizer==3.4.2
nvidia-cuda-runtime-cu12==12.6.77
sympy==1.14.0
Mako==1.3.10
Jinja2==3.1.6
orjson==3.10.18
optuna==4.3.0
pydantic_core==2.33.2
nvidia-cusparse-cu12==12.5.4.2
gitdb==4.0.12
matplotlib==3.9.4
ase==3.25.0
platformdirs==4.3.8
alembic==1.16.1
nvidia-nvjitlink-cu12==12.6.85
nvidia-cublas-cu12==12.6.4.1
annotated-types==0.7.0
networkx==3.2.1
antlr4-python3-runtime==4.9.3
nvidia-cuda-cupti-cu12==12.6.80
pillow==11.2.1
wcwidth==0.2.13
six==1.17.0
tqdm==4.67.1
torch-ema==0.3
matscipy==1.1.1
tzdata==2025.2
typing_extensions==4.13.2
sentry-sdk==2.29.1
filelock==3.18.0
python-hostlist==2.2.1
nvidia-nccl-cu12==2.26.2
docker-pycreds==0.4.0
typing-inspection==0.4.1
torchaudio==2.7.0
ConfigArgParse==1.7.1
nvidia-curand-cu12==10.3.7.77
click==8.1.8
importlib_resources==6.5.2
urllib3==2.4.0
psycopg2-binary==2.9.10
nvidia-cusolver-cu12==11.7.1.2
cycler==0.12.1
prettytable==3.16.0
eval_type_backport==0.2.2
idna==3.10
MarkupSafe==3.0.2
pyparsing==3.2.3
tomli==2.2.1
omegaconf==2.3.0
torchvision==0.22.0
packaging==25.0
psutil==7.0.0
contourpy==1.3.0
h5py==3.13.0
certifi==2025.4.26
e3nn==0.4.4
setproctitle==1.3.6
requests==2.32.3
nvidia-nvtx-cu12==12.6.77
nvidia-cusparselt-cu12==0.6.3
mpmath==1.3.0
torch==2.7.0
pandas==2.2.3
fsspec==2025.5.1
zipp==3.21.0
triton==3.3.0
SQLAlchemy==2.0.41
protobuf==6.31.0
mace-torch==0.3.13
fonttools==4.58.0
colorlog==6.9.0
pydantic==2.11.5
nvidia-cufft-cu12==11.3.0.4
nvidia-cuda-nvrtc-cu12==12.6.77
opt-einsum-fx==0.1.4
scipy==1.13.1
smmap==5.0.2
torchmetrics==1.7.1
pytz==2025.2
python-dateutil==2.9.0.post0
numpy==1.26.4
wandb==0.19.11
GitPython==3.1.44
PyYAML==6.0.2
pip==21.2.3
nvidia-cudnn-cu12==9.5.1.17
greenlet==3.2.2
lmdb==1.6.2
kiwisolver==1.4.7
nvidia-cufile-cu12==1.11.1.6
opt_einsum==3.4.0
setuptools==53.0.0
lightning-utilities==0.14.3
charset-normalizer==3.4.2
nvidia-cuda-runtime-cu12==12.6.77
sympy==1.14.0
Mako==1.3.10
Jinja2==3.1.6
orjson==3.10.18
optuna==4.3.0
pydantic_core==2.33.2
nvidia-cusparse-cu12==12.5.4.2
gitdb==4.0.12
matplotlib==3.9.4
ase==3.25.0
platformdirs==4.3.8
alembic==1.16.1
nvidia-nvjitlink-cu12==12.6.85
nvidia-cublas-cu12==12.6.4.1
annotated-types==0.7.0
networkx==3.2.1
antlr4-python3-runtime==4.9.3
nvidia-cuda-cupti-cu12==12.6.80
pillow==11.2.1
wcwidth==0.2.13
six==1.17.0
tqdm==4.67.1
torch-ema==0.3
matscipy==1.1.1
tzdata==2025.2
typing_extensions==4.13.2
sentry-sdk==2.29.1
filelock==3.18.0
python-hostlist==2.2.1
nvidia-nccl-cu12==2.26.2
docker-pycreds==0.4.0
typing-inspection==0.4.1
torchaudio==2.7.0
ConfigArgParse==1.7.1
nvidia-curand-cu12==10.3.7.77
click==8.1.8
importlib_resources==6.5.2
urllib3==2.4.0
psycopg2-binary==2.9.10
nvidia-cusolver-cu12==11.7.1.2
cycler==0.12.1
prettytable==3.16.0
eval_type_backport==0.2.2
idna==3.10
MarkupSafe==3.0.2
pyparsing==3.2.3
tomli==2.2.1
omegaconf==2.3.0
torchvision==0.22.0
packaging==25.0
psutil==7.0.0
contourpy==1.3.0
h5py==3.13.0
certifi==2025.4.26
e3nn==0.4.4
setproctitle==1.3.6
requests==2.32.3
nvidia-nvtx-cu12==12.6.77
nvidia-cusparselt-cu12==0.6.3
mpmath==1.3.0
torch==2.7.0
pandas==2.2.3
fsspec==2025.5.1
zipp==3.21.0
triton==3.3.0
SQLAlchemy==2.0.41
protobuf==6.31.0
mace-torch==0.3.13
fonttools==4.58.0
colorlog==6.9.0
pydantic==2.11.5
nvidia-cufft-cu12==11.3.0.4
nvidia-cuda-nvrtc-cu12==12.6.77
opt-einsum-fx==0.1.4
scipy==1.13.1
smmap==5.0.2
torchmetrics==1.7.1
pytz==2025.2
python-dateutil==2.9.0.post0
numpy==1.26.4
wandb==0.19.11
GitPython==3.1.44
PyYAML==6.0.2
pip==21.2.3
nvidia-cudnn-cu12==9.5.1.17
